create plr submodule
