+++ doubleml/plr/__init__.py
@@ -1,0 +1,9 @@
+"""
+The :mod:`doubleml.plr` module implements double machine learning estimates based on partially linear models.
+"""
+
+from .plr import DoubleMLPLR
+
+__all__ = [
+    "DoubleMLPLR",
+]
